added new files to manage code
