+++ messages.py
@@ -1,0 +1,28 @@
+MOSCOW_SINGLE_PLAYER_RULES = """
+Проверь своё знание Москвы!
+
+Дается неограниченное количество времени на ответ
+Можно перемещаться по улицам в любых направлениях"""
+
+RUSSIA_SINGLE_PLAYER_RULES = """
+Как хорошо ты знаешь Россию и её окрестности?
+
+Дается неограниченное количество времени на ответ
+Можно перемещаться по улицам в любых направлениях"""
+
+SPB_SINGLE_PLAYER_RULES = """
+Проверь своё знание Санкт-Петербурга!
+
+Дается неограниченное количество времени на ответ
+Можно перемещаться по улицам в любых направлениях"""
+
+HOW_TO_PLAY = """
+- Вы оказываетесь в случайной точке на карте мира
+- Можете перемещаться на панораме в любых направлениях
+
+Ваша задача: по панораме определить ваше местоположение на реальной карте мира, поставив метку на предполагаемое место вашего нахождения
+
+"""
+
+GREETING = """\nЯ - аналог игры Geoguessr в телеграме! Здесь ты можешь сыграть в отгадывание мест по всей России или по отдельным городам и посоревноваться с другими игроками
+"""
